DashClock: Fix menu overflow button on dslv
Usually menu overflow buttons on card or rows are on the top
right on google apps. Follow this UI Guideline and set it as well
to the top right.

And do not cover the whole row height as button to avoid big jumps
of the menu if it needs to show on the top of the overflow button

Change-Id: I2d78ab77249c56935624bca6d1c5949eac1f57b7